Change the LazyScopeAwareProvider in a public class
-Move the LazyScopeAwareProvider out of the DefaultConfigurer
-Make LazyScopeAwareProvider public
-Make LazyScopeAwareProvider constructor public
-Use LazyScopeAwareProvider in QuartzDeadlineManagerTest
-Use LazyScopeAwareProvider in SimpleDeadlineManagerTest

#220
