+++ core/src/test/java/org/axonframework/deadline/SimpleDeadlineManagerTest.java
@@ -16,31 +16,13 @@
 
 package org.axonframework.deadline;
 
-import org.axonframework.commandhandling.model.Repository;
 import org.axonframework.config.Configuration;
-import org.axonframework.config.SagaConfiguration;
-import org.axonframework.eventhandling.saga.AbstractSagaManager;
-import org.axonframework.eventhandling.saga.AnnotatedSagaManager;
-
-import java.util.function.Function;
+import org.axonframework.config.LazyScopeAwareProvider;
 
 public class SimpleDeadlineManagerTest extends AbstractDeadlineManagerTestSuite {
 
     @Override
     public DeadlineManager buildDeadlineManager(Configuration configuration) {
-        Repository aggregateRepository = configuration.repository(MyAggregate.class);
-        AbstractSagaManager sagaManager =
-                configuration.getModules().stream().filter(m -> m instanceof SagaConfiguration)
-                             .map(m -> (SagaConfiguration) m)
-                             .filter(sc -> sc.getSagaType().equals(MySaga.class))
-                             .map((Function<SagaConfiguration, AnnotatedSagaManager>) SagaConfiguration::getSagaManager)
-                             .findAny()
-                             .orElseThrow(() -> new IllegalStateException(String.format(
-                                     "Setup of %s test class failed, as the SagaConfiguration which is to be expected couldn't be found",
-                                     this.getClass().getSimpleName()
-                             )));
-
-        // TODO provide ScopeAwareProvider for tests
-        return new SimpleDeadlineManager(null/*aggregateRepository, sagaManager*/);
+        return new SimpleDeadlineManager(new LazyScopeAwareProvider(configuration));
     }
 }
